+++ src/Halcyon.Api/Features/EndpointTag.cs
@@ -4,6 +4,5 @@ public static class EndpointTag
 {
     public const string Account = "Account";
     public const string Profile = "Profile";
-    public const string Seed = "Seed";
     public const string Users = "Users";
 }
