--- src/Halcyon.Api/Features/Seed/SeedSettings.cs
@@ -1,8 +1,0 @@
-﻿namespace Halcyon.Api.Features.Seed;
-
-public class SeedSettings
-{
-    public static string SectionName { get; } = "Seed";
-
-    public List<SeedUser> Users { get; set; }
-}
