--- src/Halcyon.Api/Features/Seed/SeedUser.cs
@@ -1,16 +1,0 @@
-﻿namespace Halcyon.Api.Features.Seed;
-
-public class SeedUser
-{
-    public string EmailAddress { get; set; }
-
-    public string Password { get; set; }
-
-    public string FirstName { get; set; }
-
-    public string LastName { get; set; }
-
-    public DateOnly DateOfBirth { get; set; }
-
-    public List<string> Roles { get; set; }
-}
